requirements.txt
clerk-backend-api==1.6.0
fastapi==0.115.0
python-dotenv==1.0.1
python-multipart==0.0.9
pymongo==4.8.0
uvicorn==0.30.6
websockets==13.0.1

